La nota gana decimales y aprende contra qué se mide
Sobre 49 finalistas la nota entera dejaba diez nombres empatados peleando cinco plazas, y ganaba siempre el más grande. Con dos decimales eso casi desaparece. Además la nota ahora se mide contra el S&P 500 en vez de en absoluto, así que una empresa opada con retorno garantizado bajo deja de puntuar alto por pura certeza.

- El scorer y el prescore piden la nota con dos decimales reales, no redondeados a cuartos
- La nota compara contra el S&P 500, siguiendo el vocabulario del propio paper
- El oro y el petróleo entran con su precio real, no el de un ETF que a veces ni acertaba el signo del movimiento
- Volumen y float se añaden a los fundamentales, así la liquidez deja de ser una objeción sin datos detrás
- El macro deja de callarse los aranceles y empieza a comparar su previsión con la del mercado
- El constructor reintenta si la primera respuesta viene vacía o cortada, igual que ya hacía el scorer
- Todo el circuito pasa a un único modelo, más simple de razonar y sin la mezcla de dos jueces distintos
- Tres proveedores con fallos conocidos quedan excluidos del reparto de llamadas de OpenRouter
- La Sala Real tiene un botón nuevo para simular un escaneo completo sin tocar ninguna cartera
- Readme actualizado

+++ backend/app/schemas.py
@@ -14,7 +14,7 @@ class ScoreOut(BaseModel):
     created_at: datetime
     ticker: str
     sector: str
-    score: int
+    score: float          # dos decimales (ver scorer.SYSTEM); la web la redondea al pintarla
     headline: str
     report: str
     price: float | None
@@ -40,6 +40,6 @@ class WatchlistOut(BaseModel):
     model_config = ConfigDict(from_attributes=True)
 
     ticker: str
-    score: int
+    score: float
     thesis: str
     last_seen: datetime
